refactor(notifications): migrate Notification ViewModels and fix namespace alignment (#191)
- NotificationViewModel, NotificationListViewModel → partial properties
- Move NotificationListView.axaml.cs namespace to TombLauncher.Views.Notifications to match x:Class
- Fix OnPropertyChanged(nameof(ViewModels.NotificationListViewModel.Notifications)) → nameof(Notifications)
- Simplify CanClear: Enumerable.Any<T>() → Notifications.Any()
- Add using aliases for NotificationViewModel/NotificationListViewModel in affected services
- Add DependentUpon entries for notification views in .csproj
