+++ src/TombLauncher/Extensions/ViewModelsServiceCollectionExtensions.cs
@@ -1,8 +1,8 @@
 using Microsoft.Extensions.DependencyInjection;
 using TombLauncher.ViewModels;
+using TombLauncher.ViewModels.Notifications;
 using TombLauncher.ViewModels.Pages;
 using TombLauncher.ViewModels.Pages.Patchers;
-using TrxNativePatcherViewModel = TombLauncher.ViewModels.Pages.Patchers.TrxNativePatcherViewModel;
 
 namespace TombLauncher.Extensions;
