+++ src/TombLauncher/Views/Notifications/NotificationListView.axaml
@@ -3,14 +3,14 @@
              xmlns:loc="using:TombLauncher.Localization"
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-             xmlns:viewModels="clr-namespace:TombLauncher.ViewModels"
              xmlns:controls="clr-namespace:TombLauncher.Controls;assembly=TombLauncher.Controls"
              xmlns:iconPacks="https://github.com/MahApps/IconPacks.Avalonia"
+             xmlns:notifications="clr-namespace:TombLauncher.ViewModels.Notifications"
              mc:Ignorable="d" d:DesignWidth="800" d:DesignHeight="450"
-             x:Class="TombLauncher.Views.NotificationListView"
-             x:DataType="viewModels:NotificationListViewModel">
+             x:Class="TombLauncher.Views.Notifications.NotificationListView"
+             x:DataType="notifications:NotificationListViewModel">
     <Design.DataContext>
-        <viewModels:NotificationListViewModel />
+        <notifications:NotificationListViewModel />
     </Design.DataContext>
 
     <UserControl.Styles>
